Apply and optimize the program for the footages
According to the requirement of the project from the CarND, you have
to apply the program to the required three footages. Until now, the
program has already worked on three of them, with a workable color detection
to detect edge. Maybe the next step is to connect all the partial lines
togehter