Try to combine both hough transform and color
With color detection on finding path, it can precisely capture the shape of
the route, but not letting the line connected together. With hough transform,
it can create a obvious line indicating the route. If you want both the precision
of the color detection and the visibility of the hough transformation, you need to
combine both of them. However, manipulating both NDarray and matlike is not
easy, so I'm still currently working on it, or maybe just leave it still like now